+++ get_env.py
@@ -1,0 +1,3 @@
+import os
+print("EAGENT_SKILL_DATA_DIR:", os.environ.get("EAGENT_SKILL_DATA_DIR"))
+print("CURRENT_DIR:", os.getcwd())